size show for indexVC

+++ PeachPlayer/MusicModel.swift
@@ -10,7 +10,7 @@ import Foundation
 
 class MusicModel: NSObject {
     var name: String = "" //歌曲名
-    var size: Int = 0 // 歌曲大小
+    var size: Float = 0 // 歌曲大小
     var image: UIImage = UIImage() // 歌曲图片
     var data: NSData = NSData() // 歌曲内容
